@@ -10,35 +10,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.7",
-    "@angular/common": "^15.2.7",
-    "@angular/compiler": "^15.2.7",
-    "@angular/core": "^15.2.7",
-    "@angular/forms": "^15.2.7",
-    "@angular/localize": "^15.2.7",
-    "@angular/platform-browser": "^15.2.7",
-    "@angular/platform-browser-dynamic": "^15.2.7",
-    "@angular/router": "^15.2.7",
+    "@angular/animations": "^16.2.7",
+    "@angular/common": "^16.2.7",
+    "@angular/compiler": "^16.2.7",
+    "@angular/core": "^16.2.7",
+    "@angular/forms": "^16.2.7",
+    "@angular/localize": "^16.2.7",
+    "@angular/platform-browser": "^16.2.7",
+    "@angular/platform-browser-dynamic": "^16.2.7",
+    "@angular/router": "^16.2.7",
     "@fortawesome/angular-fontawesome": "^0.10.1",
     "@fortawesome/fontawesome-svg-core": "^6.0.0",
     "@fortawesome/free-brands-svg-icons": "^6.0.0",
     "@fortawesome/free-regular-svg-icons": "^6.0.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^14.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^15.1.1",
     "@popperjs/core": "^2.10.2",
     "bootstrap": "^5.2.2",
     "leaflet": "^1.9.2",
     "ng-dynamic-component": "^10.5.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.6",
-    "@angular/cli": "^15.2.6",
-    "@angular/compiler-cli": "^15.2.7",
+    "@angular-devkit/build-angular": "^16.2.4",
+    "@angular/cli": "^16.2.4",
+    "@angular/compiler-cli": "^16.2.7",
     "@types/leaflet": "^1.9.0",
     "@types/node": "^12.11.1",
-    "typescript": "~4.8.4"
+    "typescript": "~5.1.6"
   }
 }
