@@ -10,12 +10,10 @@ import { AuthService } from '../service/auth.service';
 })
 export class LoginComponent {
 
-  fg = this.fb.group({
+  fg = this.fb.nonNullable.group({
     email: [sessionStorage.getItem('email'), [Validators.required, Validators.email]],
     password: ['somestring', Validators.required],
   });
-
-  private returnUrl = this.route.snapshot.params['returnUrl'];
 
   constructor(
     private route: ActivatedRoute,
